+++ internal/fit/renderer/circle_geometry_amd64.go
@@ -12,7 +12,7 @@ var (
 func init() {
 	if cpu.X86.HasAVX2 {
 		circleSpanFloat32Backend = "avx2"
-		circleSpanFloat32Selected = circleSpanFloat32AVX2
+		circleSpanFloat32Selected = circleSpanFloat32AVX2Unchecked
 	}
 }
 
@@ -20,6 +20,10 @@ func circleSpanFloat32AVX2(centerX, radiusSquaredMinusDY float32, width int) (xS
 	if !cpu.X86.HasAVX2 {
 		return circleSpanFloat32(centerX, radiusSquaredMinusDY, width)
 	}
+	return circleSpanFloat32AVX2Unchecked(centerX, radiusSquaredMinusDY, width)
+}
+
+func circleSpanFloat32AVX2Unchecked(centerX, radiusSquaredMinusDY float32, width int) (xStart, xEnd int) {
 	cx := int(centerX + 0.5)
 	return circleSpanFloat32AVX2Kernel(centerX, radiusSquaredMinusDY, float32(cx), width)
 }